fitur login hardcoded

+++ src/main/java/com/anam/elibrary/controller/UserController.java
@@ -10,6 +10,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
+import javax.servlet.http.HttpSession;
+
 @Controller
 public class UserController {
 
@@ -17,40 +19,58 @@ public class UserController {
     private UserService userService;
 
     @GetMapping("/users")
-    public String users(Model model) {
+    public String users(Model model, HttpSession session) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         model.addAttribute("titlePage", "USERS");
         model.addAttribute("users", userService.findAll());
         return "dashboard/users";
     }
 
     @GetMapping("/users/create")
-    public String usersCreate(Model model) {
+    public String usersCreate(Model model, HttpSession session) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         model.addAttribute("titlePage", "USERS CREATE");
         model.addAttribute("user", new User());
         return "dashboard/users-create";
     }
 
     @PostMapping("/users/create")
-    public String usersCreateSubmit(@ModelAttribute(value = "user") User user) {
+    public String usersCreateSubmit(HttpSession session, @ModelAttribute(value = "user") User user) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         userService.save(user);
         return "redirect:/users";
     }
 
     @GetMapping("/users/edit/{id}")
-    public String userEdit(Model model, @PathVariable("id") int id) {
+    public String userEdit(Model model, HttpSession session, @PathVariable("id") int id) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         model.addAttribute("titlePage", "USERS EDIT");
         model.addAttribute("user", userService.findById(id));
         return "dashboard/users-edit";
     }
 
     @PostMapping("/users/edit")
-    public String usersEditSubmit(@ModelAttribute(value = "user") User user) {
+    public String usersEditSubmit(HttpSession session, @ModelAttribute(value = "user") User user) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         userService.update(user);
         return "redirect:/users";
     }
 
     @GetMapping("/users/delete/{id}")
-    public String usersDeleteSubmit(@PathVariable("id") int id) {
+    public String usersDeleteSubmit(HttpSession session, @PathVariable("id") int id) {
+        if (session.getAttribute("username") == null) {
+            return "redirect:/login";
+        }
         userService.deleteById(id);
         return "redirect:/users";
     }
